backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-cors-headers==3.13.0
django-filter==22.1
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
python-dotenv==1.0.0
sqlparse==0.4.2
requests==2.26.0

